requirements.txt
pdfplumber
python-docx
sentence-transformers
faiss-cpu
numpy
